Refactor: 임시글 하위 컴포넌트 이동
